Refactor CollectibleResource actor structure
Moved CollectibleResource actor files to a new directory structure and updated includes. Added Damageable interface implementation, new mesh component, and TryApplyDamage logic to the actor. Renamed header and source files to reflect new organization.
